% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TabelaOrdinal.R
\name{TabelaOrdinal}
\alias{TabelaOrdinal}
\title{Cria tabela de frequências para uma variável de tipo ordinal}
\usage{
TabelaOrdinal(y)
}
\description{
Cria tabela de frequência absoluta e frequência relativa com totais por coluna para uma variável de tipo ordinal
}
\details{
@param y Vetor fator ordenado

@return Matriz

@examples
 aux <- mtcars$cyl
 TabelaOrdinal(aux)

@export
}
