% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TabelaQuantitativa.R
\name{TabelaQuantitativa}
\alias{TabelaQuantitativa}
\title{Cria tabela de frequências para uma variável de tipo quantitativa}
\usage{
TabelaQuantitativa(y)
}
\description{
Cria tabela de frequência absoluta,frequência relativa, frequência absoluta acumulada e frequência relativa acumulada com totais por coluna para uma variável de tipo quantitativa
}
\details{
@param y Vetor numérico

@return Matriz

@examples
 aux <- mtcars$wt
 TabelaQuantitativa(aux)

@export
}
